refactor: 💡 move table specific objects around
